Added abritrary item spawning
Type "p" in the console to give yourself 5 prism stones. I still don't really know what param3 of the original function does since it seems to always be 1, but param1 seems to be some kind of event or location ID. I'll need to test more items to see if it's OK/safe to assume a 0 value means the function is being called from inside the house.